remove legacy expression resolver and codegen for alps and elasticdl
