+++ pkg/sql/submitter.go
@@ -23,7 +23,6 @@ import (
 	"path"
 	"sync"
 
-	"sqlflow.org/sqlflow/pkg/parser"
 	pb "sqlflow.org/sqlflow/pkg/proto"
 	"sqlflow.org/sqlflow/pkg/sql/codegen/tensorflow"
 	"sqlflow.org/sqlflow/pkg/sql/codegen/xgboost"
@@ -33,9 +32,8 @@ import (
 var envSubmitter = os.Getenv("SQLFLOW_submitter")
 
 var submitterRegistry = map[string](Submitter){
-	"default":   &defaultSubmitter{},
-	"alps":      &alpsSubmitter{&defaultSubmitter{}},
-	"elasticdl": &elasticdlSubmitter{&defaultSubmitter{}},
+	"default": &defaultSubmitter{},
+	// add submitters like alps, elasticdl
 }
 
 func submitter() Submitter {
@@ -100,9 +98,6 @@ type defaultSubmitter struct {
 	Cwd      string
 	Session  *pb.Session
 }
-
-type elasticdlSubmitter struct{ *defaultSubmitter }
-type alpsSubmitter struct{ *defaultSubmitter }
 
 func (s *defaultSubmitter) Setup(w *PipeWriter, db *DB, modelDir string, session *pb.Session) error {
 	// cwd is used to store train scripts and save output models.
@@ -208,39 +203,3 @@ func (s *defaultSubmitter) ExecuteAnalyze(cl *ir.AnalyzeStmt) error {
 }
 func (s *defaultSubmitter) Teardown()                   { os.RemoveAll(s.Cwd) }
 func (s *defaultSubmitter) GetTrainStmtFromModel() bool { return true }
-
-func (s *elasticdlSubmitter) ExecuteTrain(cl *ir.TrainStmt) error {
-	// TODO(typhoonzero): remove below twice parse when all submitters moved to IR.
-	pr, e := parser.LegacyParse(cl.OriginalSQL)
-	if e != nil {
-		return e
-	}
-	return elasticDLTrain(s.Writer, pr, s.Db, s.Cwd, s.Session)
-}
-
-func (s *elasticdlSubmitter) ExecutePredict(cl *ir.PredictStmt) error {
-	// TODO(typhoonzero): remove below twice parse when all submitters moved to IR.
-	pr, e := parser.LegacyParse(cl.OriginalSQL)
-	if e != nil {
-		return e
-	}
-	return elasticDLPredict(s.Writer, pr, s.Db, s.Cwd, s.Session)
-}
-
-func (s *alpsSubmitter) ExecuteTrain(cl *ir.TrainStmt) error {
-	// TODO(typhoonzero): remove below twice parse when all submitters moved to IR.
-	pr, e := parser.LegacyParse(cl.OriginalSQL)
-	if e != nil {
-		return e
-	}
-	return alpsTrain(s.Writer, pr, s.Db, s.Cwd, s.Session)
-}
-
-func (s *alpsSubmitter) ExecutePredict(cl *ir.PredictStmt) error {
-	// TODO(typhoonzero): remove below twice parse when all submitters moved to IR.
-	pr, e := parser.LegacyParse(cl.OriginalSQL)
-	if e != nil {
-		return e
-	}
-	return alpsPred(s.Writer, pr, s.Db, s.Cwd, s.Session)
-}
